move ActionLink to action_link.go
